Create alphabetconvertable module since the same methods were used for both encrypt and decrypt classes
