+++ lib/decryptmessage.rb
@@ -1,28 +1,7 @@
+require_relative 'alphabetconvertable'
+
 class DecryptMessage
-  def alphabet
-    ("a".."z").to_a << " "
-  end
-
-  def alpha_with_index
-    index_placement = alphabet.map.with_index do |letter, position|
-      {letter => position}
-    end
-    index_placement.reduce({}, :merge)
-  end
-
-  def convert(letter)
-    convert = []
-    letter_into_array = []
-    letter_into_array << letter
-    letter_into_array.each do |char|
-      if alpha_with_index.include?(char)
-        convert.push alpha_with_index[char]
-      else
-        convert.push(char)
-      end
-    end
-    convert
-  end
+  include Aplhabetcovertable
 
   def decrypt_single_letter(letter, shift)
     convert(letter).map do |number|
