+++ lib/enigma.rb
@@ -1,5 +1,5 @@
 class Enigma
-
+  
   def initialize
     @cipherkey = CipherKey.new
     @offset = Offset.new
